fakefullscreen: fix for mis-sized fullscreen applications
An example problematic situation would be to open a YouTube video
in Google Chrome and making that video fullscreen. Without this fix
the window will go into fullscreen within its tiled dimension, but
the size of the video will match that of the screen (and it will
be cropped).

With this fix the window will be resized which nudges the
application to adjust the fullscreen size accordingly.

(cherry picked from commit b4ae62d21a38b475d34c8a1b3a051e41ccda92dd)